@@ -51,63 +51,53 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
-    "@angular/cdk": "~13.3.0",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/elements": "~13.3.0",
-    "@angular/forms": "~13.3.0",
-    "@angular/material": "~13.3.0",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
-    "@senzing/rest-api-client-ng": "^5.1.0-beta.1",
-    "acorn": "^8.6.0",
-    "ajv": "^6.12.6",
+    "@angular/animations": "^15.0.4",
+    "@angular/cdk": "^15.0.3",
+    "@angular/common": "^15.0.4",
+    "@angular/compiler": "^15.0.4",
+    "@angular/core": "^15.0.4",
+    "@angular/elements": "^15.0.4",
+    "@angular/forms": "^15.0.4",
+    "@angular/material": "^15.0.3",
+    "@angular/platform-browser": "^15.0.4",
+    "@angular/platform-browser-dynamic": "^15.0.4",
+    "@angular/router": "^15.0.4",
+    "@senzing/rest-api-client-ng": "^6.0.0-alpha.1",
     "d3": "^5.9.1",
     "document-register-element": "^1.14.10",
-    "html2canvas": "^1.0.0-rc.7",
-    "jspdf": "^2.4.0",
-    "ngx-json-viewer": "^3.0.2",
-    "rxjs": "~7.4.0",
+    "ngx-json-viewer": "^3.2.1",
+    "rxjs": "~7.5.0",
     "sass": "^1.54.5",
     "tslib": "^2.3.0",
-    "zone.js": "~0.11.7"
+    "zone.js": "~0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.0",
-    "@angular-devkit/core": "^13.3.0",
-    "@angular-devkit/schematics": "^13.3.0",
-    "@angular/cli": "~13.3.0",
-    "@angular/compiler-cli": "~13.3.0",
-    "@angular/language-service": "^13.3.0",
-    "@compodoc/compodoc": "^1.1.18",
+    "@angular-devkit/build-angular": "^15.0.4",
+    "@angular/cli": "~15.0.4",
+    "@angular/compiler-cli": "^15.0.4",
+    "@angular/language-service": "^15.0.4",
+    "@compodoc/compodoc": "^1.1.19",
     "@types/d3": "^5.9.1",
-    "@types/jasmine": "~3.5.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/jspdf": "^1.3.0",
-    "@types/node": "^12.11.1",
-    "angular-in-memory-web-api": "^0.13.0",
+    "@types/jasmine": "~4.3.1",
     "cheerio": "^1.0.0-rc.12",
     "concat": "^1.0.3",
     "concurrently": "^6.4.0",
     "fs-extra": "^9.0.1",
-    "jasmine-core": "~4.3.0",
+    "jasmine-core": "~4.5.0",
+    "jasmine-spec-reporter": "~7.0.0",
     "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.1.0",
+    "karma-chrome-launcher": "~3.1.1",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~1.7.0",
-    "karma-summary-reporter": "~3.1.1",
-    "ng-packagr": "^13.3.0",
+    "karma-jasmine-html-reporter": "~2.0.0",
+    "ng-packagr": "^15.0.3",
     "ngx-webstorage-service": "^5.0.0",
     "protractor": "~7.0.0",
     "rimraf": "^3.0.2",
     "sass-burger": "^1.3.1",
     "scss-bundle": "^3.1.2",
     "ts-node": "~10.9.1",
-    "typescript": "~4.4.4",
+    "typescript": "~4.8.4",
     "wait-on": "^6.0.1"
   },
   "optionalDependencies": {
