@@ -1,6 +1,6 @@
 {
   "name": "@senzing/sdk-components-ng",
-  "version": "5.1.1-beta.2",
+  "version": "6.0.0-ng15",
   "private": false,
   "keywords" :["Angular","Library","Senzing"],
   "license" : "SEE LICENSE IN LICENSE",
@@ -11,18 +11,15 @@
   "homepage" :"https://senzing.github.io/sdk-components-ng/",
   "dependencies": {
     "@types/d3": "^5.9.1",
-    "@types/jspdf": "^1.3.0",
     "d3": "^5.9.1",
-    "html2canvas": "^1.0.0-rc.3",
-    "jspdf": "^2.3.1",
-    "ngx-json-viewer": "^2.4.0",
+    "ngx-json-viewer": "^3.2.1",
     "tslib": "^2.3.0"
   },
   "peerDependencies": {
-    "@angular/cdk": "^13.3.0",
-    "@angular/common": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/material": "~13.3.0",
-    "@senzing/rest-api-client-ng": "^5.0.0"
+    "@angular/cdk": "^15.0.3",
+    "@angular/common": "^15.0.4",
+    "@angular/core": "^15.0.4",
+    "@angular/material": "^15.0.3",
+    "@senzing/rest-api-client-ng": "^6.0.0"
   }
 }
